Refactor ResultsContainer Rename ResultsContainer to LinksContainer Refactor link logic to Link component
