+++ src/App.js
@@ -1,7 +1,7 @@
 import { useQuery } from "@apollo/client";
 import { useEffect, useState } from "react";
 import Form from "./components/Form";
-import ResultsContainer from "./components/ResultsContainer";
+import LinksContainer from "./components/LinksContainer";
 import { GET_ALL_LINKS } from "./graphQL/queries";
 import "./styles.css";
 
@@ -17,9 +17,9 @@ export default function App() {
   if (error) return `Error! ${error}`;
 
   return (
-    <div className="App bg-gray-100 min-h-screen xl:max-w-screen-xl mx-auto">
+    <div className="App bg-gray-100 min-h-screen xl:max-w-screen-2xl mx-auto">
       <Form refetch={refetch} />
-      <ResultsContainer links={allLinks} />
+      <LinksContainer links={allLinks} />
     </div>
   );
 }
